requirements_agents.txt
# Agent System Dependencies

# Google Gemini API Client
google-generativeai>=0.3.0

# Environment Variable Management
python-dotenv>=1.0.0

# Schema Validation
pydantic>=2.0.0

# Existing project dependencies (if not already installed)
# Add any additional dependencies needed for optimizer integration
